+++ packages/cli/src/controller/codegen-controller.test.ts
@@ -12,43 +12,38 @@ jest.setTimeout(30000);
 
 describe('Codegen can generate schema', () => {
   afterEach(async () => {
-    await promisify(rimraf)(path.join(__dirname, '../../test/schemaTest1/src'));
-    await promisify(rimraf)(path.join(__dirname, '../../test/schemaTest2/src'));
-    await promisify(rimraf)(path.join(__dirname, '../../test/schemaTest3/src'));
-    await promisify(rimraf)(path.join(__dirname, '../../test/schemaTest4/src'));
-    await promisify(rimraf)(path.join(__dirname, '../../test/schemaTest5/src'));
-    await promisify(rimraf)(path.join(__dirname, '../../test/schemaTest6/src'));
+    await promisify(rimraf)(path.join(__dirname, '../../test/schemaTest/src'));
   });
 
   it('codegen with correct schema should pass', async () => {
-    const projectPath = path.join(__dirname, '../../test/schemaTest1');
-    await codegen(projectPath);
+    const projectPath = path.join(__dirname, '../../test/schemaTest');
+    await expect(codegen(projectPath)).resolves.not.toThrow();
   });
 
   it('codegen with incorrect schema field should fail', async () => {
-    const projectPath = path.join(__dirname, '../../test/schemaTest2');
-    await expect(codegen(projectPath)).rejects.toThrow(/is not an valid type/);
+    const projectPath = path.join(__dirname, '../../test/schemaTest');
+    await expect(codegen(projectPath, ['project-bad-schema.yaml'])).rejects.toThrow(/is not an valid type/);
   });
   it('codegen with entities that uses reserved names should throw', async () => {
-    const projectPath = path.join(__dirname, '../../test/schemaTest3');
-    await expect(codegen(projectPath)).rejects.toThrow(
+    const projectPath = path.join(__dirname, '../../test/schemaTest');
+    await expect(codegen(projectPath, ['project-bad-entity.yaml'])).rejects.toThrow(
       'EntityName: exampleEntityFilter cannot end with reservedKey: filter'
     );
   });
   it('Codegen should be able to generate ABIs from template datasources', async () => {
-    const projectPath = path.join(__dirname, '../../test/schemaTest4');
-    await codegen(projectPath);
+    const projectPath = path.join(__dirname, '../../test/schemaTest');
+    await codegen(projectPath, ['project-templates-abi.yaml']);
     await expect(
       fs.promises.readFile(`${projectPath}/src/types/abi-interfaces/Erc721.ts`, 'utf8')
     ).resolves.toBeTruthy();
   });
 
   it('Should not fail, if ds does not have any assets', async () => {
-    const projectPath = path.join(__dirname, '../../test/schemaTest5');
-    await expect(codegen(projectPath)).resolves.not.toThrow();
+    const projectPath = path.join(__dirname, '../../test/schemaTest');
+    await expect(codegen(projectPath, ['project-no-assets.yaml'])).resolves.not.toThrow();
   });
   it('Codegen should be able to generate ABIs from customName datasources', async () => {
-    const projectPath = path.join(__dirname, '../../test/schemaTest6');
+    const projectPath = path.join(__dirname, '../../test/schemaTest');
     await codegen(projectPath);
     await expect(
       fs.promises.readFile(`${projectPath}/src/types/abi-interfaces/Erc721.ts`, 'utf8')
@@ -56,7 +51,7 @@ describe('Codegen can generate schema', () => {
   });
 
   it('Should clean out existing types directory', async () => {
-    const projectPath = path.join(__dirname, '../../test/schemaTest6');
+    const projectPath = path.join(__dirname, '../../test/schemaTest');
     await codegen(projectPath);
     await codegen(projectPath, ['project-no-abi.yaml']);
 
@@ -64,7 +59,7 @@ describe('Codegen can generate schema', () => {
     await expect(fs.promises.readFile(`${projectPath}/src/types/abi-interfaces/Erc721.ts`, 'utf8')).rejects.toThrow();
   });
   it('should generate contracts on different glob paths', async () => {
-    const projectPath = path.join(__dirname, '../../test/schemaTest6');
+    const projectPath = path.join(__dirname, '../../test/schemaTest');
     await codegen(projectPath, ['typechain-test.yaml']);
 
     await expect(
@@ -80,7 +75,7 @@ describe('Codegen can generate schema', () => {
     ).resolves.toBeTruthy();
   });
   it('Should not generate ABI for non evm ds', async () => {
-    const projectPath = path.join(__dirname, '../../test/schemaTest6');
+    const projectPath = path.join(__dirname, '../../test/schemaTest');
     await codegen(projectPath, ['non-evm-project.yaml']);
     expect(fs.existsSync(`${projectPath}/src/types/abi-interfaces/`)).toBeFalsy();
   });