+++ packages/cli/src/controller/codegen-controller.ts
@@ -1,7 +1,6 @@
 // Copyright 2020-2023 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
-import fs from 'fs';
 import path from 'path';
 import {DEFAULT_MANIFEST, getManifestPath, getSchemaPath, loadFromJsonOrYaml} from '@subql/common';
 import {
@@ -20,6 +19,7 @@ import {
   isRuntimeDs as isRuntimeEthereumDs,
   RuntimeDatasourceTemplate as EthereumDsTemplate,
   CustomDatasourceTemplate as EthereumCustomDsTemplate,
+  generateAbis,
 } from '@subql/common-ethereum';
 import {
   isCustomDs as isCustomNearDs,
@@ -51,7 +51,6 @@ import {
   getAllEnums,
 } from '@subql/utils';
 import {upperFirst, uniq, uniqBy} from 'lodash';
-import {runTypeChain, glob, parseContractPath} from 'typechain';
 import {renderTemplate, prepareDirPath} from '../utils';
 
 type TemplateKind =
@@ -72,22 +71,11 @@ const MODEL_TEMPLATE_PATH = path.resolve(__dirname, '../template/model.ts.ejs');
 const MODELS_INDEX_TEMPLATE_PATH = path.resolve(__dirname, '../template/models-index.ts.ejs');
 const TYPES_INDEX_TEMPLATE_PATH = path.resolve(__dirname, '../template/types-index.ts.ejs');
 const INTERFACE_TEMPLATE_PATH = path.resolve(__dirname, '../template/interface.ts.ejs');
-const ABI_INTERFACE_TEMPLATE_PATH = path.resolve(__dirname, '../template/abi-interface.ts.ejs');
 const ENUM_TEMPLATE_PATH = path.resolve(__dirname, '../template/enum.ts.ejs');
 const DYNAMIC_DATASOURCE_TEMPLATE_PATH = path.resolve(__dirname, '../template/datasource-templates.ts.ejs');
 const TYPE_ROOT_DIR = 'src/types';
 const MODEL_ROOT_DIR = 'src/types/models';
-const ABI_INTERFACES_ROOT_DIR = 'src/types/abi-interfaces';
-const CONTRACTS_DIR = 'src/types/contracts'; //generated
-const TYPECHAIN_TARGET = 'ethers-v5';
-
 const RESERVED_KEYS = ['filter', 'filters'];
-const CUSTOM_EVM_HANDLERS = [
-  'cosmos/EthermintEvm',
-  'substrate/FrontierEvm',
-  'substrate/AcalaEvm',
-  'substrate/Moonbeam',
-];
 
 const exportTypes = {
   models: false,
@@ -161,156 +149,6 @@ export async function generateEnums(projectPath: string, schema: string): Promis
       throw new Error(`When render enums having problems.`);
     }
   }
-}
-
-export interface abiRenderProps {
-  name: string;
-  events: string[];
-  functions: {typeName: string; functionName: string}[];
-}
-export interface abiInterface {
-  name: string;
-  type: 'event' | 'function';
-  inputs: {
-    internalType: string;
-    name: string;
-    type: string;
-  }[];
-}
-function validateCustomDs(d: DatasourceKind) {
-  return CUSTOM_EVM_HANDLERS.includes(d.kind);
-}
-
-export async function generateAbis(datasources: DatasourceKind[], projectPath: string): Promise<void> {
-  const sortedAssets = new Map<string, string>();
-  datasources.map((d) => {
-    if (!d?.assets) {
-      return;
-    }
-    if (isRuntimeEthereumDs(d) || isCustomEthereumDs(d) || validateCustomDs(d)) {
-      Object.entries(d.assets).map(([name, value]) => {
-        const filePath = path.join(projectPath, value.file);
-        if (!fs.existsSync(filePath)) {
-          throw new Error(`Error: Asset ${name}, file ${value.file} does not exist`);
-        }
-        // We use actual abi file name instead on name provided in assets
-        // This is aligning with files in './ethers-contracts'
-        sortedAssets.set(parseContractPath(filePath).name, value.file);
-      });
-    }
-  });
-  if (sortedAssets.size === 0) {
-    return prepareDirPath(path.join(projectPath, ABI_INTERFACES_ROOT_DIR), false);
-  }
-
-  await prepareDirPath(path.join(projectPath, ABI_INTERFACES_ROOT_DIR), true);
-  try {
-    const allFiles = glob(projectPath, [...sortedAssets.values()]);
-    // Typechain generate interfaces under CONTRACTS_DIR
-    // Run typechain for individual paths, if provided glob paths are the same,
-    // it will generate incorrect file structures, and fail to generate contracts for certain abis
-    await Promise.all(
-      allFiles.map((file) =>
-        runTypeChain({
-          cwd: projectPath,
-          filesToProcess: [file],
-          allFiles: [file],
-          outDir: CONTRACTS_DIR,
-          target: TYPECHAIN_TARGET,
-        })
-      )
-    );
-    // Iterate here as we have to make sure type chain generated successful,
-    // also avoid duplicate generate same abi interfaces
-    const renderAbiJobs = processAbis(sortedAssets, projectPath, loadFromJsonOrYamlWrapper);
-    await Promise.all(
-      renderAbiJobs.map((renderProps) => {
-        console.log(`* Abi Interface ${renderProps.name} generated`);
-        return renderTemplate(
-          ABI_INTERFACE_TEMPLATE_PATH,
-          path.join(projectPath, ABI_INTERFACES_ROOT_DIR, `${renderProps.name}.ts`),
-          {
-            props: {abi: renderProps},
-            helper: {upperFirst},
-          }
-        );
-      })
-    );
-  } catch (e) {
-    console.error(`! Unable to generate abi interface. ${e.message}`);
-  }
-}
-
-function loadFromJsonOrYamlWrapper(filePath: string): abiInterface[] | {abi: abiInterface[]} {
-  return loadFromJsonOrYaml(filePath) as abiInterface[] | {abi: abiInterface[]};
-}
-
-export function processAbis(
-  sortedAssets: Map<string, string>,
-  projectPath: string,
-  loadReadAbi: (filePath: string) => abiInterface[] | {abi: abiInterface[]}
-): abiRenderProps[] {
-  const renderInterfaceJobs: abiRenderProps[] = [];
-  sortedAssets.forEach((value, key) => {
-    const renderProps: abiRenderProps = {name: key, events: [], functions: []};
-    const readAbi = loadReadAbi(path.join(projectPath, value));
-    // We need to use for loop instead of map, due to events/function name could be duplicate,
-    // because they have different input, and following ether typegen rules, name also changed
-    // we need to find duplicates, and update its name rather than just unify them.
-
-    let abiArray: abiInterface[] = [];
-
-    if (!Array.isArray(readAbi)) {
-      if (!readAbi.abi) {
-        throw new Error(`Provided ABI is not a valid ABI or Artifact`);
-      }
-      abiArray = readAbi.abi;
-    } else {
-      abiArray = readAbi;
-    }
-
-    if (!abiArray.length) {
-      throw new Error(`Invalid abi is provided at asset: ${key}`);
-    }
-
-    const duplicateEventNames = abiArray
-      .filter((abiObject) => abiObject.type === 'event')
-      .map((obj) => obj.name)
-      .filter((name, index, arr) => arr.indexOf(name) !== index);
-    const duplicateFunctionNames = abiArray
-      .filter((abiObject) => abiObject.type === 'function')
-      .map((obj) => obj.name)
-      .filter((name, index, arr) => arr.indexOf(name) !== index);
-    abiArray.map((abiObject) => {
-      if (abiObject.type === 'function') {
-        let typeName = abiObject.name;
-        let functionName = abiObject.name;
-        if (duplicateFunctionNames.includes(abiObject.name)) {
-          functionName = `${abiObject.name}(${abiObject.inputs.map((obj) => obj.type.toLowerCase()).join(',')})`;
-          typeName = joinInputAbiName(abiObject);
-        }
-        renderProps.functions.push({typeName, functionName});
-      }
-      if (abiObject.type === 'event') {
-        let name = abiObject.name;
-        if (duplicateEventNames.includes(abiObject.name)) {
-          name = joinInputAbiName(abiObject);
-        }
-        renderProps.events.push(name);
-      }
-    });
-    // avoid empty json
-    if (!!renderProps.events || !!renderProps.functions) {
-      renderInterfaceJobs.push(renderProps);
-    }
-  });
-  return renderInterfaceJobs;
-}
-
-export function joinInputAbiName(abiObject: abiInterface): string {
-  // example: "TextChanged_bytes32_string_string_string_Event", Event name/Function type name will be joined in ejs
-  const inputToSnake = abiObject.inputs.map((obj) => obj.type.replace(/\[\]/g, '_arr').toLowerCase()).join('_');
-  return `${abiObject.name}_${inputToSnake}_`;
 }
 
 export function processFields(
@@ -439,7 +277,7 @@ export async function codegen(projectPath: string, fileNames: string[] = [DEFAUL
   }
   await generateCosmwasm(datasources, projectPath, prepareDirPath, upperFirst, renderTemplate);
 
-  await generateAbis(datasources, projectPath);
+  await generateAbis(datasources, projectPath, prepareDirPath, upperFirst, renderTemplate);
 
   if (exportTypes.interfaces || exportTypes.models || exportTypes.enums || exportTypes.datasources) {
     try {