+++ packages/cli/src/controller/generate-controller.ts
@@ -10,13 +10,13 @@ import {
   EthereumHandlerKind,
   EthereumTransactionFilter,
   SubqlRuntimeHandler,
+  parseContractPath,
 } from '@subql/common-ethereum';
 import {SubqlRuntimeDatasource as EthereumDs, EthereumLogFilter} from '@subql/types-ethereum';
 import chalk from 'chalk';
 import {Interface} from 'ethers/lib/utils';
 import * as inquirer from 'inquirer';
 import {upperFirst, difference, pickBy} from 'lodash';
-import {parseContractPath} from 'typechain';
 import {Document, parseDocument, YAMLSeq} from 'yaml';
 import {SelectedMethod, UserInput} from '../commands/codegen/generate';
 import {renderTemplate, resolveToAbsolutePath} from '../utils';